added development and release mode for gamelift deployment

+++ deployBuildToGameLift.sh
@@ -41,12 +41,15 @@ for i in `seq 0 $SIZE`; do
 	FLEET_ID=$(echo $EXISTING_FLEETS | jq ".FleetAttributes[$i].FleetId" | sed 's/"//g')
 	INSTANCES=$(aws gamelift describe-instances --fleet-id $FLEET_ID | jq '.Instances | length')
 	STATUS=$(echo "$EXISTING_FLEETS" | jq ".FleetAttributes[$i].Status" | sed 's/"//g')
-	if [ "$INSTANCES" = "0" -a "$STATUS" = "ACTIVE" ]; then
-		FLEET_ID=$(echo "$EXISTING_FLEETS" | jq ".FleetAttributes[$i].FleetId" | sed 's/"//g')
-		delete_game_session_queues "$FLEET_ARN"
-		aws gamelift delete-fleet --fleet-id "$FLEET_ID"
-		BUILD_ID="$(echo "$EXISTING_FLEETS" | jq ".FleetAttributes[$i].BuildId" | sed 's/"//g')"
-		aws gamelift delete-build --build-id "$BUILD_ID"
+	NAME=$(echo "EXISTING_FLEETS" | jq ".FleetAttributes[$i].Name" | sed 's/"//g')
+	if [ ! -z "$(echo "$NAME" | grep "$3")" ]; then
+		if [ "$INSTANCES" = "0" -a "$STATUS" = "ACTIVE" ]; then
+			FLEET_ID=$(echo "$EXISTING_FLEETS" | jq ".FleetAttributes[$i].FleetId" | sed 's/"//g')
+			delete_game_session_queues "$FLEET_ARN"
+			aws gamelift delete-fleet --fleet-id "$FLEET_ID"
+			BUILD_ID="$(echo "$EXISTING_FLEETS" | jq ".FleetAttributes[$i].BuildId" | sed 's/"//g')"
+			aws gamelift delete-build --build-id "$BUILD_ID"
+		fi
 	fi
 done
 
@@ -77,8 +80,8 @@ rm -rf ./downloaded-builds/LinuxServer
 
 wget -m -nH --cut-dirs=2 -P ./downloaded-builds/LinuxServer ftp://jenkins:$2@ftp.valhalla-game.com/$1/LinuxServer
 
- curl https://raw.githubusercontent.com/saiaku-gaming/jenkins-common/master/gamelift-install.sh > downloaded-builds/LinuxServer/install.sh
- chmod +x ./downloaded-builds/LinuxServer/install.sh
+curl https://raw.githubusercontent.com/saiaku-gaming/jenkins-common/master/gamelift-install.sh > downloaded-builds/LinuxServer/install.sh
+chmod +x ./downloaded-builds/LinuxServer/install.sh
 
 #pre create an empty saved logs file so that graylog can monitor the folder on system start.
 mkdir -p ./downloaded-builds/LinuxServer/valhalla/Saved/Logs
@@ -86,7 +89,7 @@ mkdir -p ./downloaded-builds/LinuxServer/valhalla/Saved/Logs
 chmod +x ./downloaded-builds/LinuxServer/valhalla/Binaries/Linux/valhallaServer
 
 echo "Uploading build..."
-OUTPUT="$(aws gamelift upload-build --name "$1 Build" --build-version $1 --build-root ./downloaded-builds/LinuxServer --operating-system AMAZON_LINUX --region "eu-central-1")"
+OUTPUT="$(aws gamelift upload-build --name "$3 $1 Build" --build-version $1 --build-root ./downloaded-builds/LinuxServer --operating-system AMAZON_LINUX --region "eu-central-1")"
 BUILD_ID="$(echo $OUTPUT | tail -n 1 | sed 's/^.*Build ID: //')"
 echo "Build uploaded with id: $BUILD_ID"
 COUNT=0
@@ -100,7 +103,7 @@ while [ "$READY" != "READY" ] && [ $COUNT -ne 60 ]; do
 done
 
 echo "Creating fleet..."
-FLEET_RESPONSE=$(aws gamelift create-fleet --name "$3 Fleet" --build-id "$BUILD_ID" --ec2-instance-type "c4.large" --ec2-inbound-permissions '[{"FromPort": 7777,"ToPort": 7787,"IpRange": "0.0.0.0/0","Protocol": "UDP"},{"FromPort": 8990,"ToPort": 9000,"IpRange": "0.0.0.0/0","Protocol": "TCP"}]' --runtime-configuration '{"ServerProcesses": [{"LaunchPath": "/local/game/valhalla/Binaries/Linux/valhallaServer", "Parameters": "-Log -GameLift", "ConcurrentExecutions": 10}], "MaxConcurrentGameSessionActivations": 10, "GameSessionActivationTimeoutSeconds": 1}')
+FLEET_RESPONSE=$(aws gamelift create-fleet --name "$3 $1 Fleet" --build-id "$BUILD_ID" --ec2-instance-type "c4.large" --ec2-inbound-permissions '[{"FromPort": 7777,"ToPort": 7787,"IpRange": "0.0.0.0/0","Protocol": "UDP"},{"FromPort": 8990,"ToPort": 9000,"IpRange": "0.0.0.0/0","Protocol": "TCP"}]' --runtime-configuration '{"ServerProcesses": [{"LaunchPath": "/local/game/valhalla/Binaries/Linux/valhallaServer", "Parameters": "-Log -GameLift", "ConcurrentExecutions": 10}], "MaxConcurrentGameSessionActivations": 10, "GameSessionActivationTimeoutSeconds": 1}')
 
 FLEET_ARN=$(echo $FLEET_RESPONSE | jq .FleetAttributes.FleetArn | sed s/\"//g)
 FLEET_ID=$(echo $FLEET_RESPONSE | jq .FleetAttributes.FleetId | sed s/\"//g)
@@ -110,7 +113,7 @@ echo "Adding scaling policies"
 
 aws gamelift put-scaling-policy --name "Scale up" --fleet-id "$FLEET_ID" --scaling-adjustment "1" --scaling-adjustment-type "ExactCapacity" --threshold "1" --comparison-operator "GreaterThanOrEqualToThreshold" --evaluation-periods "1" --metric-name "QueueDepth"
 
-aws gamelift put-scaling-policy --name "Scale down" --fleet-id "$FLEET_ID" --scaling-adjustment "0" --scaling-adjustment-type "ExactCapacity" --threshold "1" --comparison-operator "LessThanThreshold" --evaluation-periods "120" --metric-name "ActiveGameSessions"
+aws gamelift put-scaling-policy --name "Scale down" --fleet-id "$FLEET_ID" --scaling-adjustment "0" --scaling-adjustment-type "ExactCapacity" --threshold "1" --comparison-operator "LessThanThreshold" --evaluation-periods "60" --metric-name "ActiveGameSessions"
 
 #RESPONSE=$(aws gamelift describe-game-session-queues)
 #SIZE=$(echo $RESPONSE | jq '.GameSessionQueues | length')
